# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Whiskey Golf Fantasy Golf API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /tournaments:
    get:
      operationId: listTournaments
      tags: [tournaments]
      summary: List all tournaments
      responses:
        "200":
          description: List of tournaments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tournament"
    post:
      operationId: createTournament
      tags: [tournaments]
      summary: Create a new tournament
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTournamentBody"
      responses:
        "201":
          description: Tournament created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tournament"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /tournaments/{id}:
    get:
      operationId: getTournament
      tags: [tournaments]
      summary: Get tournament by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tournament with config, teams, and field
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TournamentDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateTournament
      tags: [tournaments]
      summary: Update tournament
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTournamentBody"
      responses:
        "200":
          description: Updated tournament
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tournament"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /tournaments/{id}/config:
    get:
      operationId: getTournamentConfig
      tags: [tournaments]
      summary: Get tournament config
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tournament config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TournamentConfig"
    patch:
      operationId: updateTournamentConfig
      tags: [tournaments]
      summary: Update tournament config
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTournamentConfigBody"
      responses:
        "200":
          description: Updated config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TournamentConfig"
  /tournaments/{id}/position-points:
    get:
      operationId: getPositionPoints
      tags: [tournaments]
      summary: Get position points table for a tournament
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Position points array
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PositionPoints"
    post:
      operationId: setPositionPoints
      tags: [tournaments]
      summary: Set (replace) position points table
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetPositionPointsBody"
      responses:
        "200":
          description: Updated position points
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PositionPoints"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: string
        createdAt:
          type: string
      required:
        - id
        - username
        - displayName
        - createdAt
    Tournament:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        courseName:
          type: ["string", "null"]
        startDate:
          type: string
        endDate:
          type: string
        status:
          type: string
          enum: [draft, live, completed]
        notes:
          type: ["string", "null"]
        winnerId:
          type: ["integer", "null"]
        createdAt:
          type: string
        completedAt:
          type: ["string", "null"]
      required:
        - id
        - name
        - startDate
        - endDate
        - status
        - createdAt
    TournamentDetail:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        courseName:
          type: ["string", "null"]
        startDate:
          type: string
        endDate:
          type: string
        status:
          type: string
          enum: [draft, live, completed]
        notes:
          type: ["string", "null"]
        winnerId:
          type: ["integer", "null"]
        createdAt:
          type: string
        completedAt:
          type: ["string", "null"]
        config:
          $ref: "#/components/schemas/TournamentConfig"
        golfers:
          type: array
          items:
            $ref: "#/components/schemas/TournamentGolfer"
        teams:
          type: array
          items:
            $ref: "#/components/schemas/FantasyTeam"
      required:
        - id
        - name
        - startDate
        - endDate
        - status
        - createdAt
        - config
        - golfers
        - teams
    CreateTournamentBody:
      type: object
      properties:
        name:
          type: string
        courseName:
          type: ["string", "null"]
        startDate:
          type: string
        endDate:
          type: string
        notes:
          type: ["string", "null"]
        config:
          $ref: "#/components/schemas/UpdateTournamentConfigBody"
      required:
        - name
        - startDate
        - endDate
    UpdateTournamentBody:
      type: object
      properties:
        name:
          type: string
        courseName:
          type: ["string", "null"]
        startDate:
          type: string
        endDate:
          type: string
        status:
          type: string
          enum: [draft, live, completed]
        notes:
          type: ["string", "null"]
    TournamentConfig:
      type: object
      properties:
        id:
          type: integer
        tournamentId:
          type: integer
        draftType:
          type: string
          enum: [alternate, snake]
        salaryCap:
          type: integer
        rosterSize:
          type: integer
        captainMultiplier:
          type: number
        birdiePoints:
          type: number
        eaglePoints:
          type: number
        bogeyPenalty:
          type: number
        missedCutPenalty:
          type: number
        replacementTopRankLockout:
          type: integer
        requireAmerican:
          type: boolean
        requireEuropean:
          type: boolean
        requireRow:
          type: boolean
        requireOutsideTop30:
          type: boolean
        salaryMin:
          type: integer
        salaryMax:
          type: integer
        scoringPlaces:
          type: integer
        firstPlacePoints:
          type: integer
      required:
        - id
        - tournamentId
        - draftType
        - salaryCap
        - rosterSize
        - captainMultiplier
        - birdiePoints
        - eaglePoints
        - bogeyPenalty
        - missedCutPenalty
        - replacementTopRankLockout
        - requireAmerican
        - requireEuropean
        - requireRow
        - requireOutsideTop30
        - salaryMin
        - salaryMax
        - scoringPlaces
        - firstPlacePoints
    UpdateTournamentConfigBody:
      type: object
      properties:
        draftType:
          type: string
          enum: [alternate, snake]
        salaryCap:
          type: integer
        rosterSize:
          type: integer
        captainMultiplier:
          type: number
        birdiePoints:
          type: number
        eaglePoints:
          type: number
        bogeyPenalty:
          type: number
        missedCutPenalty:
          type: number
        replacementTopRankLockout:
          type: integer
        requireAmerican:
          type: boolean
        requireEuropean:
          type: boolean
        requireRow:
          type: boolean
        requireOutsideTop30:
          type: boolean
        salaryMin:
          type: integer
        salaryMax:
          type: integer
        scoringPlaces:
          type: integer
        firstPlacePoints:
          type: integer
    PositionPoints:
      type: object
      properties:
        position:
          type: integer
        points:
          type: number
      required:
        - position
        - points
    SetPositionPointsBody:
      type: object
      properties:
        points:
          type: array
          items:
            $ref: "#/components/schemas/PositionPoints"
      required:
        - points
    TournamentGolfer:
      type: object
      properties:
        id:
          type: integer
        tournamentId:
          type: integer
        golferName:
          type: string
        nationality:
          type: string
        region:
          type: string
          enum: [EU, US, ROW]
        avgOdds:
          type: ["number", "null"]
        worldRanking:
          type: ["integer", "null"]
        marketRank:
          type: ["integer", "null"]
        salary:
          type: ["integer", "null"]
        outsideTop30:
          type: boolean
        eligibleForReplacement:
          type: boolean
      required:
        - id
        - tournamentId
        - golferName
        - nationality
        - region
        - outsideTop30
        - eligibleForReplacement
...
    FantasyTeam:
      type: object
      properties:
        id:
          type: integer
        tournamentId:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        userDisplayName:
          type: string
        totalSalary:
          type: ["integer", "null"]
        totalScore:
          type: ["number", "null"]
        isWinner:
          type: boolean
        picks:
          type: array
          items:
            $ref: "#/components/schemas/TeamPick"
      required:
        - id
        - tournamentId
        - userId
        - userName
        - userDisplayName
        - isWinner
        - picks
    TeamPick:
      type: object
      properties:
        id:
          type: integer
        fantasyTeamId:
          type: integer
        golferPoolId:
          type: integer
        golferName:
          type: string
        region:
          type: string
        salary:
          type: ["integer", "null"]
        slotNumber:
          type: integer
        isCaptain:
          type: boolean
        missedCut:
          type: boolean
        replacementGolferPoolId:
          type: ["integer", "null"]
        replacementGolferName:
          type: ["string", "null"]
        slotScorePreMultiplier:
          type: ["number", "null"]
        slotScorePostMultiplier:
          type: ["number", "null"]
      required:
        - id
        - fantasyTeamId
        - golferPoolId
        - golferName
        - region
        - slotNumber
        - isCaptain
        - missedCut
        - isWinner
...