# lib/api-zod/src/generated/types/setPositionPointsBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Whiskey Golf Fantasy Golf API
 * OpenAPI spec version: 0.1.0
 */
import type { PositionPoints } from "./positionPoints";

export interface SetPositionPointsBody {
  points: PositionPoints[];
}