# lib/api-zod/src/generated/types/tournamentConfig.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Whiskey Golf Fantasy Golf API
 * OpenAPI spec version: 0.1.0
 */
import type { TournamentConfigDraftType } from "./tournamentConfigDraftType";

export interface TournamentConfig {
  id: number;
  tournamentId: number;
  draftType: TournamentConfigDraftType;
  salaryCap: number;
  rosterSize: number;
  captainMultiplier: number;
  birdiePoints: number;
  eaglePoints: number;
  bogeyPenalty: number;
  missedCutPenalty: number;
  replacementTopRankLockout: number;
  requireAmerican: boolean;
  requireEuropean: boolean;
  requireRow: boolean;
  requireOutsideTop30: boolean;
  salaryMin: number;
  salaryMax: number;
  scoringPlaces: number;
  firstPlacePoints: number;
}